requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# NLP - Entity Extraction Only
spacy==3.7.2

# Intent Classification
scikit-learn==1.4.0
joblib==1.3.2
numpy==1.26.3

# Rule Engine
json-logic-qubit==0.9.1

# JSON Schema Validation
jsonschema==4.21.1

# MCP Protocol
httpx==0.26.0
websockets==12.0

# Logging and Monitoring
structlog==24.1.0
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
orjson==3.9.12
